+++ autumn/migrations/20260917120000_create_money_ledger/down.sql
@@ -1,3 +1,5 @@
+DROP TRIGGER IF EXISTS _autumn_money_postings_no_truncate ON _autumn_money_postings;
+DROP TRIGGER IF EXISTS _autumn_money_transactions_no_truncate ON _autumn_money_transactions;
 DROP TRIGGER IF EXISTS _autumn_money_postings_append_only ON _autumn_money_postings;
 DROP TRIGGER IF EXISTS _autumn_money_transactions_append_only ON _autumn_money_transactions;
 DROP FUNCTION IF EXISTS _autumn_money_append_only();